fix(rank): 순위 조회 쿼리 바인딩 에러 수정

+++ app/api/mathrank-rank-read-api/src/main/java/kr/co/mathrank/rank/read/RankResponse.java
@@ -4,7 +4,7 @@ import kr.co.mathrank.domain.rank.dto.RankQueryResult;
 import kr.co.mathrank.domain.rank.entity.Tier;
 
 public record RankResponse(
-	Integer rank,
+	Long rank,
 	Tier tier,
 	Long score,
 	Long totalUserCount
