+++ domain/mathrank-rank-domain/src/main/java/kr/co/mathrank/domain/rank/dto/RankQueryResult.java
@@ -3,12 +3,12 @@ package kr.co.mathrank.domain.rank.dto;
 import kr.co.mathrank.domain.rank.entity.Tier;
 
 public record RankQueryResult(
-	Integer rank,
+	Long rank,
 	Tier tier,
 	Long score,
 	Long totalUserCount
 ) {
-	public static RankQueryResult of(final int rank, final Tier tier, final Long score, final long totalUserCount) {
+	public static RankQueryResult of(final Long rank, final Tier tier, final Long score, final long totalUserCount) {
 		return new RankQueryResult(rank, tier, score, totalUserCount);
 	}
 }
